+++ core/compiler/optimization.cpp
@@ -924,7 +924,7 @@ void ItermediateOptimizer::ApplyReduction(IntermediateInstruction* test, Interme
 
   std::deque<IntermediateInstruction*> rewrite_instrs;
   if(shift) {
-    rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)shift));
+    rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)shift));
     // exclude literal
     if(working_stack.front()->GetType() != LOAD_INT_LIT) {
       rewrite_instrs.push_back(working_stack.front());
@@ -935,10 +935,10 @@ void ItermediateOptimizer::ApplyReduction(IntermediateInstruction* test, Interme
     working_stack.pop_front();
     // shift left or right
     if(instr->GetType() == MUL_INT) {
-      rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, SHL_INT, (int)shift));
+      rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, SHL_INT, (long)shift));
     } 
     else {
-      rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, SHR_INT, (int)shift));
+      rewrite_instrs.push_back(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, SHR_INT, (long)shift));
     }
   }
   
@@ -1296,7 +1296,7 @@ IntermediateBlock* ItermediateOptimizer::ConstantProp(IntermediateBlock* inputs)
       if(instr->GetOperand2() == LOCL) {
         std::unordered_map<int, PropValue>::iterator result = value_prop_map.find(instr->GetOperand());
         if(result != value_prop_map.end()) {
-          outputs->AddInstruction(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)result->second.int_value));
+          outputs->AddInstruction(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)result->second.int_value));
         }
         else {
           outputs->AddInstruction(instr);
@@ -1443,49 +1443,49 @@ void ItermediateOptimizer::CalculateIntFold(IntermediateInstruction* instr, std:
     switch(instr->GetType()) {
     case ADD_INT: {
       INT_VALUE value = left->GetOperand() + right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case SUB_INT: {
       INT_VALUE value = left->GetOperand() - right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case MUL_INT: {
       INT_VALUE value = left->GetOperand() * right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case DIV_INT: {
       INT_VALUE value = left->GetOperand() / right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case MOD_INT: {
       INT_VALUE value = left->GetOperand() % right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
       
     case BIT_AND_INT: {
       INT_VALUE value = left->GetOperand() & right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case BIT_OR_INT: {
       INT_VALUE value = left->GetOperand() | right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
 
     case BIT_XOR_INT: {
       INT_VALUE value = left->GetOperand() ^ right->GetOperand();
-      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (int)value));
+      working_stack.push_front(IntermediateFactory::Instance()->MakeInstruction(cur_line_num, LOAD_INT_LIT, (long)value));
     }
       break;
